Clean up committee related helper funcs
