+++ beacon-chain/cache/BUILD.bazel
@@ -44,6 +44,7 @@ go_library(
         "//shared/copyutil:go_default_library",
         "//shared/featureconfig:go_default_library",
         "//shared/hashutil:go_default_library",
+        "//shared/mathutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/rand:go_default_library",
         "//shared/sliceutil:go_default_library",
