+++ beacon-chain/core/helpers/BUILD.bazel
@@ -4,14 +4,15 @@ go_library(
     name = "go_default_library",
     srcs = [
         "attestation.go",
+        "beacon_committee.go",
         "block.go",
-        "committee.go",
         "genesis.go",
         "randao.go",
         "rewards_penalties.go",
         "shuffle.go",
         "signing_root.go",
         "slot_epoch.go",
+        "sync_committee.go",
         "validators.go",
         "weak_subjectivity.go",
     ],
@@ -60,13 +61,14 @@ go_test(
     size = "medium",
     srcs = [
         "attestation_test.go",
+        "beacon_committee_test.go",
         "block_test.go",
-        "committee_test.go",
         "randao_test.go",
         "rewards_penalties_test.go",
         "shuffle_test.go",
         "signing_root_test.go",
         "slot_epoch_test.go",
+        "sync_committee_test.go",
         "validators_test.go",
         "weak_subjectivity_test.go",
     ],
